Version 5.1 - Логика работы с приоритетными задачами перенесена в InMemoryTaskManager
