+++ src/history/HistoryManager.java
@@ -10,14 +10,4 @@ public interface HistoryManager {
     void remove(int id);
 
     List<Task> getHistory();
-
-    List<Task> getPrioritizedTasks();
-
-    void addToPrioritizedTasks(Task task);
-
-    void removePrioritizedTask(Task oldTask);
-
-    boolean isIntersectionTasks(Task newTask);
-
-    boolean isIntersectionTasks(Task newTask, int idAddedTask);
 }
